Abdullah Mohammad Jaid_A23MJ4001/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.2,<25.0
pypdf>=4.0,<6.0
python-docx>=1.1,<2.0
requests>=2.31,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.0,<12.0
pytesseract>=0.3,<0.4
